Backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2
python-dotenv
datetime